feat: lesson 16

+++ app/Http/Controllers/Admin/ProductController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\ProductRequest;
 use App\Models\Category;
 use App\Models\Product;
+
 //use Illuminate\Http\Request;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
@@ -82,12 +83,22 @@ class ProductController extends Controller
      */
     public function update(ProductRequest $request, Product $product)
     {
-        Storage::delete($product->image);
-        $path = $request->file('image')->store('products');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')) {
+            Storage::delete($product->image);
+            $params['image'] = $request->file('image')->store('products');
+        }
+
         $product->update($params);
         return redirect()->route('products.index');
+
+//        Storage::delete($product->image);
+//        $path = $request->file('image')->store('products');
+//        $params = $request->all();
+//        $params['image'] = $path;
+//        $product->update($params);
+//        return redirect()->route('products.index');
     }
 
     /**
